속도 측정 단위를 nanos 로 변경

+++ Source/ARX/include/ARX/ARTracker2d.h
@@ -100,7 +100,7 @@ private:
     bool m_running;
     bool unloadTwoDData();
     bool loadTwoDData(std::vector<ARTrackable *>& trackables);
-    double now_ms();
+    long now_ms();
 };
 
 #endif // HAVE_2D
